Subscription management: Migrate  to client package
