+++ client/landing/subscriptions/subscription-manager/styles.scss
@@ -1,109 +1,46 @@
 @import "@automattic/color-studio/dist/color-variables";
 @import "@wordpress/base-styles/breakpoints";
 
-.section-nav.subscription-manager-tab-switcher {
-	box-shadow: none;
-	margin-bottom: 32px;
+.subscription-manager__container {
+	max-width: 1440px;
+	padding: 48px 62px;
 
-	.section-nav-tabs {
-		display: flex;
+	.formatted-header {
+		margin: 0;
 
-		.section-nav-tabs__list {
-			box-shadow: inset 0 -1px 0 rgba($studio-black, 0.05);
-			gap: 16px;
-			flex-basis: fit-content;
+		&__title {
+			font-size: $font-headline-medium !important;
+			line-height: $font-headline-large;
+			color: $studio-gray-100;
+			margin-bottom: 16px;
+		}
 
-			.section-nav-tab {
-				border: none;
+		&__subtitle {
+			font-size: $font-body !important;
+			line-height: $font-title-medium;
+			color: $studio-gray-60;
+			margin-bottom: 30px;
 
-				&.is-selected {
-					.section-nav-tab__link {
-						border-bottom-color: $studio-gray-100;
-					}
-
-					.section-nav-tab__text {
-						color: $studio-gray-100;
-					}
-				}
-
-				&:hover:not(.is-selected) .section-nav-tab__link {
-					border-bottom-color: $studio-gray-40;
-				}
-
-				.section-nav-tab__link {
-					border-bottom: solid 2px transparent;
-					color: $studio-gray-60;
-					padding: 8px 0;
-					transition: border-bottom-color ease-in-out 0.1s, color ease-in-out 0.1s;
-
-					&:hover {
-						background-color: transparent;
-					}
-				}
-
-				.section-nav-tab__text {
-					color: $studio-gray-60;
-					font-size: 1rem;
-					line-height: 24px;
-				}
-
-				.count {
-					background: $studio-gray-5;
-					border: none;
-					border-radius: 2px;
-					color: $studio-gray-80;
-					font-size: 0.75rem;
-					line-height: 20px;
-					margin-left: 10px;
-					padding: 0 8px;
-				}
+			& .email-address {
+				font-weight: 500;
 			}
 		}
 	}
 }
 
+
 @media screen and ( max-width: $break-medium ) {
-	.section-nav.subscription-manager-tab-switcher {
-		margin-bottom: 22px;
+	.subscription-manager__container {
+		padding: 30px 24px;
+
+		.formatted-header {
+			&__title {
+				font-size: rem(28px) !important;
+				line-height: $font-headline-small;
+				margin-bottom: 8px;
+			}
+		}
+
 	}
 }
 
-@media screen and ( max-width: $break-medium ) {
-	.section-nav.subscription-manager-tab-switcher {
-		.section-nav__mobile-header {
-			display: none;
-		}
-
-		.section-nav-group {
-			display: flex;
-		}
-
-		.section-nav__panel {
-			display: flex;
-			align-items: center;
-		}
-
-		.section-nav-tab__text {
-			display: inline;
-			width: auto;
-			text-align: center;
-			flex-grow: 0;
-			font-weight: 400;
-		}
-
-		.section-nav-tab__link {
-			display: block;
-			font-weight: 400;
-			background-color: transparent;
-		}
-
-		.section-nav-tabs {
-			flex-grow: 1;
-			flex-shrink: 0;
-		}
-
-		.section-nav-tabs__list {
-			display: flex;
-		}
-	}
-}
