+++ client/landing/subscriptions/subscription-manager/subscription-manager.tsx
@@ -1,9 +1,14 @@
 import { Reader } from '@automattic/data-stores';
-import SubscriptionManager from '@automattic/subscription-manager';
+import { UniversalNavbarHeader } from '@automattic/wpcom-template-parts';
 import { useTranslate } from 'i18n-calypso';
 import { useMemo } from 'react';
+import DocumentHead from 'calypso/components/data/document-head';
+import FormattedHeader from 'calypso/components/formatted-header';
+import Main from 'calypso/components/main';
+import { useSubheaderText } from '../hooks';
 import { Settings } from '../tab-views';
 import { TabsSwitcher } from '../tabs-switcher';
+import './styles.scss';
 
 const SitesView = () => <span>Sites View</span>;
 const CommentsView = () => <span>Comments View</span>;
@@ -35,9 +40,23 @@ const SubscriptionManagementPage = () => {
 	);
 
 	return (
-		<SubscriptionManager>
-			<TabsSwitcher baseRoute="subscriptions" tabs={ tabs } />
-		</SubscriptionManager>
+		<>
+			<UniversalNavbarHeader
+				className="subscription-manager-header"
+				variant="minimal"
+				isLoggedIn={ false }
+			/>
+			<Main className="subscription-manager__container">
+				<DocumentHead title="Subscriptions" />
+				<FormattedHeader
+					brandFont
+					headerText={ translate( 'Subscription management' ) }
+					subHeaderText={ useSubheaderText() }
+					align="left"
+				/>
+				<TabsSwitcher baseRoute="subscriptions" tabs={ tabs } />
+			</Main>
+		</>
 	);
 };
