--- packages/subscription-manager/README.md
@@ -1,4 +1,0 @@
-# Subscription Manager
-This package creates the subscription management views that are displayed as:
-1. Part of Reader for wpcom logged-in users
-2. Externally facing single page app for users who use their e-mail address for subscribing to wpcom websites, but they do not have a wpcom account
