--- packages/subscription-manager/package.json
@@ -1,51 +1,0 @@
-{
-	"name": "@automattic/subscription-manager",
-	"version": "0.0.1",
-	"description": "Subnscription Manager",
-	"homepage": "https://github.com/Automattic/wp-calypso",
-	"license": "GPL-2.0-or-later",
-	"author": "Automattic Inc.",
-	"main": "dist/cjs/index.js",
-	"module": "dist/esm/index.js",
-	"calypso:src": "src/index.tsx",
-	"sideEffects": [
-		"*.css",
-		"*.scss"
-	],
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/Automattic/wp-calypso.git",
-		"directory": "packages/subscription-manager"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"bugs": "https://github.com/Automattic/wp-calypso/issues",
-	"types": "dist/types",
-	"scripts": {
-		"clean": "tsc --build ./tsconfig.json ./tsconfig-cjs.json --clean && rm -rf dist",
-		"build": "tsc --build ./tsconfig.json ./tsconfig-cjs.json && run -T copy-assets",
-		"prepack": "yarn run clean && yarn run build",
-		"watch": "tsc --build ./tsconfig.json --watch"
-	},
-	"dependencies": {
-		"@automattic/data-stores": "workspace:^",
-		"@automattic/typography": "workspace:^",
-		"@automattic/wpcom-template-parts": "workspace:^",
-		"i18n-calypso": "workspace:^",
-		"react": "^17.0.2",
-		"react-router-dom": "^5.1.2"
-	},
-	"devDependencies": {
-		"@automattic/calypso-typescript-config": "workspace:^",
-		"@automattic/typography": "workspace:^",
-		"typescript": "^4.5.5"
-	},
-	"peerDependencies": {
-		"@wordpress/data": "^7.6.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"reakit-utils": "^0.15.1",
-		"redux": "^4.1.2"
-	}
-}
