--- packages/subscription-manager/src/components/SubscriptionManagerContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as SubscriptionManagerContainer } from './SubscriptionManagerContainer';
