--- packages/subscription-manager/src/index.tsx
@@ -1,1 +1,0 @@
-export { SubscriptionManagerContainer as default } from './components/SubscriptionManagerContainer';
