--- packages/subscription-manager/tsconfig-cjs.json
@@ -1,7 +1,0 @@
-{
-	"extends": "./tsconfig",
-	"compilerOptions": {
-		"module": "commonjs",
-		"outDir": "dist/cjs"
-	}
-}
