--- packages/subscription-manager/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"extends": "@automattic/calypso-typescript-config/ts-package.json",
-	"compilerOptions": {
-		"declarationDir": "dist/types",
-		"outDir": "dist/esm",
-		"rootDir": "src"
-	},
-	"include": [ "src" ],
-	"exclude": [ "**/test/*" ],
-	"references": [
-		{ "path": "../wpcom-template-parts" },
-		{ "path": "../data-stores" },
-		{ "path": "../components" }
-	]
-}
