+++ packages/tsconfig.json
@@ -48,7 +48,6 @@
 		{ "path": "./sites" },
 		{ "path": "./state-utils" },
 		{ "path": "./subscriber" },
-		{ "path": "./subscription-manager" },
 		{ "path": "./tree-select" },
 		{ "path": "./viewport" },
 		{ "path": "./viewport-react" },
